refactor test_lifecycle
