+++ test/src/realtime/test_realtime.cpp
@@ -30,125 +30,6 @@ void test_tournament();
 void test_rpc();
 void test_rt_party();
 
-void test_rt_quickdestroy()
-{
-    NTest test(__func__);
-    test.runTest([&test]() {
-        test.listener.setErrorCallback([&test](const NRtError&) {
-            test.stopTest(false);
-        });
-
-        test.listener.setConnectCallback([&test]() {
-            test.rtClient->disconnect();
-            test.stopTest(true);
-        });
-        test.rtClient->connect(test.session, true, test.protocol);
-    });
-}
-
-void test_rt_rapiddisconnect()
-{
-    NTest test(__func__);
-    test.setTestTimeoutMs(1000);
-    bool hadConnectCallback = false;
-
-    test.runTest([&test, &hadConnectCallback]() {
-
-        // if no callbacks were fired, it is a success
-        test.onTimeoutCb = []() { return true; };
-
-        test.listener.setConnectCallback([&hadConnectCallback]() {
-            NLOG_INFO("TEST: Connect");
-            hadConnectCallback = true;
-        });
-        test.listener.setDisconnectCallback([&test, &hadConnectCallback](const NRtClientDisconnectInfo& /*info*/) {
-            NLOG_INFO("TEST: Disconnect");
-            // both connect and disconnect callback fired and disconnect fired after connect
-            test.stopTest(hadConnectCallback);
-        });
-        test.rtClient->connect(test.session, true, test.protocol);
-        test.rtClient->disconnect();
-    });
-}
-
-void test_rt_reconnect()
-{
-    NTest test(__func__);
-    bool connectedOnce = false;
-    test.onRtConnect = [&test, &connectedOnce](){
-        NLOG_INFO("TEST: connected");
-        test.listener.setDisconnectCallback([&test](const NRtClientDisconnectInfo& /*info*/){
-            NLOG_INFO("TEST: disconnected, connecting again");
-            test.rtClient->connect(test.session, true, test.protocol);
-        });
-        test.rtClient->disconnect();
-        if(connectedOnce) {
-            NLOG_INFO("TEST: Stopping test with success");
-            test.stopTest(true); // connect after disconnect succeeded
-        } else {
-            connectedOnce = true;
-        }
-    };
-    test.runTest();
-}
-
-void test_rt_heartbeat()
-{
-    NTest test(__func__);
-    test.setRtStopTestOnDisconnect(false);
-    test.setTestTimeoutMs(100000);
-    int i = 0;
-    test.onRtConnect = [&test, &i](){
-        NLOG_INFO("CONNECTED");
-        // Fail heartbeat immediately on the second tick
-        // Adjusting heartbeat timeout like below is not really representative case of a network failue,
-        // because server sends close frame and it gets delivered. To test it for real, comment line below
-        // and use `pkill -STOP nakama` (linux) or sysinternal process explorer to suspend nakama process
-        // making it unresponsive. You'll need to "unfreeze" it again with `pkill -CONT` mid test, so that
-        // it can connect again.
-        test.rtClient->setHeartbeatIntervalMs(0);
-        test.listener.setDisconnectCallback([&test, &i](const NRtClientDisconnectInfo& info){
-            if (info.code != NRtClientDisconnectInfo::Code::HEARTBEAT_FAILURE) {
-                NLOG_INFO("Unexpected heartbeat failure disconnect code, wanted=4000 , got=" + std::to_string(info.code));
-                test.stopTest(false);
-            }
-
-            if (i++ > 0) {
-                test.stopTest(true);
-                return;
-            }
-            // test that heartbeats continue to work after reconnect
-            test.rtClient->connect(test.session, true, test.protocol);
-        });
-    };
-    test.runTest();
-}
-
-// optional test -- pass --socket.pong_wait_ms 5000 to Nakama so that it disconnects rtclient after sleep
-// in order to test how it will react.
-void test_rt_remote_disconnect()
-{
-    NTest test(__func__);
-    // we want to test behavior of socket after disconnect.
-    test.setRtStopTestOnDisconnect(false);
-    test.setTestTimeoutMs(15000);
-    test.onRtConnect = [&test]()
-    {
-        test.rtClient->createMatch([&test](const Nakama::NMatch& match)
-        {
-            test.setRtTickPaused(true);
-            std::this_thread::sleep_for(std::chrono::milliseconds(10000));
-
-            test.setRtTickPaused(false);
-
-            std::this_thread::sleep_for(std::chrono::milliseconds(1000));
-
-            test.stopTest(true);
-        });
-    };
-
-    test.runTest();
-}
 
 void run_realtime_tests()
 {
